+++ FastMarching3D_ROS/src/frontier_planner_node.cpp
@@ -19,7 +19,7 @@
 // Cost map global
 MapGrid3D<double> speedMap;
 pcl::PointCloud<pcl::PointXYZI>::Ptr speedMapCloud(new pcl::PointCloud<pcl::PointXYZI>);
-float voxelSize = 0.2;
+double voxelSize = 0.2;
 bool mapUpdated = false;
 bool firstMapReceived = false;
 double speedMax = 10.0;
@@ -94,7 +94,7 @@ void CallbackSpeedMap(const sensor_msgs::PointCloud2::ConstPtr msg)
 
   pcl::fromROSMsg(*msg, *speedMapCloud);
   // Get bounds
-  float boundsMin[3], boundsMax[3];
+  double boundsMin[3], boundsMax[3];
   GetPointCloudBounds(speedMapCloud, boundsMin, boundsMax);
 
   // Copy cloud to speedMap data structure
@@ -302,7 +302,7 @@ int main(int argc, char **argv)
   n.param("frontier_planner/num_agents", numGoals, (int)1);
   float goalSeparationDistance, turnRate;
   n.param("frontier_planner/goal_separation_distance", goalSeparationDistance, (float)5.0);
-  n.param("frontier_planner/voxel_size", voxelSize, (float)0.2);
+  n.param("frontier_planner/voxel_size", voxelSize, (double)0.2);
   n.param("frontier_planner/speed_max", speedMax, (double)10.0);
   n.param("frontier_planner/speed_safe", speedSafe, (double)0.4);
   n.param("frontier_planner/turn_rate", turnRate, (float)15.0); // deg/s
